==> 01_Seznam_cerpacich.ipynb <==
import pandas as pd

seznam_cerpacich = pd.read_csv("Seznam_cerpacich_final.csv", sep= ";", encoding = "utf-8")

seznam_cerpacich 


seznam_cerpacich[seznam_cerpacich["ID"].isin([288, 337, 1930, 2290, 3187])]

seznam_cerpacich = seznam_cerpacich[~seznam_cerpacich["ID"].isin([288, 337, 1930, 2290, 3187])]

seznam_cerpacich[seznam_cerpacich["ID"].isin([288, 337, 1930, 2290, 3187])]

seznam_cerpacich.info()

seznam_cerpacich.columns

seznam_cerpacich.drop("LAU 1", axis=1, inplace=True)

seznam_cerpacich.rename(columns = {"Jiná":"Ostatní mimo elektřiny"}, inplace = True)

cols = list(seznam_cerpacich.columns)
cols.insert(1, cols.pop(cols.index("Datum uvedení do provozu")))
seznam_cerpacich = seznam_cerpacich[cols]

seznam_cerpacich

seznam_cerpacich.to_csv("01_vystup_seznam_cerpacich_stanic.csv", sep=";", index=False, encoding="utf-8")
