==> 02_Verejne_dostupne.ipynb <==
import pandas as pd

verejne_dostupne_CZ = pd.read_csv("Verejne_dostupne_stanice_CZ.csv", sep=",", skiprows=5, encoding="utf-8", quotechar='"')

# Zobraz všechny sloupce, ne jen zkrácené
pd.set_option('display.max_columns', None)

verejne_dostupne_CZ.head(5)

verejne_dostupne_CZ.shape

verejne_dostupne_CZ.drop("Unnamed: 0", axis=1, inplace=True) 

verejne_dostupne_CZ.rename(columns={
"ulice":"Ulice",
"pozice (pořadí stanice na stejné adrese)":"Pořadí stanice na stejné adrese",
"obec":"Obec",
"provozní doba":"Provozní doba",
"ID CPO (identifi-kátor provozo-vatele dobíjecí stanice)":"ID CPO - identifikátor provozovatele",
"celkový nominální max. výkon stanice (kW)":"Celkový max. výkon stanice(kW)",
"pocet dobíjecích bodů":"Počet dobíjecích bodů",
"typ dobíjení (AC, DC, UFC)":"Typ dobíjení (1.konektor)",
"nominální maximální výkon konektoru (kW)":"Max. výkon 1.konektoru(kW)",
"standard konektoru":"Standard 1.konektoru",
"způsob připojení ":"Způsob připojení 1",
"vyloučený souběh s konektory":"Vyloučený souběh s konektory",
"typ dobíjení (AC, DC, UFC).1":"Typ dobíjení (2.konektor)",
"nominální maximální výkon konektoru (kW).1":"Max. výkon 2.konektoru(kW)",
"standard konektoru.1":"Standard 2.konektoru",
"způsob připojení .1":"Způsob připojení 2",
"vyloučený souběh s konektory č..1":"Vyloučený souběh s konektory",
"typ dobíjení (AC, DC, UFC).2":"Typ dobíjení (3.konektor)",
"nominální maximální výkon konektoru (kW).2":"Max. výkon 3.konektoru(kW)",
"standard konektoru.2":"Standard 3.konektoru",
"způsob připojení .2":"Způsob připojení 3",
"vyloučený souběh s konektory č..2":"Vyloučený souběh s konektory"
}, inplace=True)

# stanice má 1 dobíjecí bod, typ dobíjení AC (střídavý), pomalejší), výkon 22,2kW, Typ 2 je evropský standard
#vyloučený souběh NaN, protože je jen jeden konektor

# stanice má 2 dobíjecí body, typ dobíjení DC (stejnosměrný) pokud je zapojen jen jeden konektor 
# tak dostane plných 50kW, konektor(1) CCS2 rychlonabíjecí na kabel (nemusíme mít vlastní)
# pokud se zapnout dva konektory, tak si ten výkon musí mezi sebe rozdělit, takže 25kW a 25kW

#UFC(ultra-fast)


cols = list(verejne_dostupne_CZ.columns)

cols.insert(0, cols.pop(cols.index("Datum uvedení do provozu")))

verejne_dostupne_CZ = verejne_dostupne_CZ[cols]

verejne_dostupne_CZ["Datum uvedení do provozu"] = pd.to_datetime(
    verejne_dostupne_CZ["Datum uvedení do provozu"],
    format="%m/%d/%Y"
)

verejne_dostupne_CZ["Datum uvedení do provozu"] = verejne_dostupne_CZ["Datum uvedení do provozu"].dt.strftime("%d.%m.%Y")

verejne_dostupne_CZ.insert(0, "ID", range(1, len(verejne_dostupne_CZ) + 1))

verejne_dostupne_CZ

verejne_dostupne_CZ.to_csv("02_vystup_verejne_dostupne.csv", sep=";", index=False, encoding="utf-8")
