==> 03_Spotreba_vkrajich.ipynb <==
import pandas as pd

spotreba_vkrajich = pd.read_csv("Spotreba_vkrajich_sektory.csv", sep=",", encoding = "utf-8")

spotreba_vkrajich.head(20)

spotreba_vkrajich["Jednotka"].unique()

spotreba_vkrajich.rename(columns = {"Spotřeba":"Spotřeba (GWh)"}, inplace = True)

spotreba_vkrajich.drop("Jednotka", axis=1, inplace=True)

spotreba_vkrajich

spotreba_vkrajich.to_csv("03_vystup_spotreba_kraje.csv", sep=";", index=False, encoding="utf-8")
