==> 04_podil_OZ.ipynb <==
import pandas as pd

podil_EU = pd.read_csv("podil_OZ_naspotrebe_vEU.csv", sep=",", encoding = "utf-8")

podil_EU

podil_EU.shape

podil_EU.info()

podil_EU.rename(columns = {"Podíl obnovitelných":"Podíl obnovitelných zdrojů (%)"}, inplace = True)

podil_EU.drop("Jednotka", axis=1, inplace=True)

podil_EU["Země"].value_counts()

podil_EU

podil_EU.to_csv("04_vystup_podil_OZ.csv", sep=";", index=False, encoding="utf-8")
