==> 06_RSV_bez_chybnych_radku.ipynb <==
#### z RSV_bez_chybnych_radku jsme identifikovali pouze řádky, kde je Kategorie vozidla "M1"

import pandas as pd

chunksize = 100_000  # velikost jednoho kusu

# připravíme prázdný DataFrame
final = pd.DataFrame()

typy_osobaku = [
    "M1"
]

osobaky = []
# postupně čteme velký soubor po kouskách
for chunk in pd.read_csv(
    "06_RSV_bez_chybnych_radku.csv",
    sep=",",
    on_bad_lines="skip",
    quoting=0,
    encoding="utf-8",
    low_memory=False,
    chunksize=chunksize):  # <- tady říkáme "čti po 100 000 řádcích"
        
    # filtrujeme řádky
    osobaky = chunk[chunk["Kategorie vozidla"].isin(typy_osobaku)]
    
    # připojíme vybraný kus do finálního DataFrame
    final = pd.concat([final, osobaky], ignore_index=True)

# na konci uložíme celé zpracované CSV
final.to_csv("06_RSV_bez_chybnych_radku_M1.csv", index=False, encoding="utf-8")
