==> 06_RSV_palivo_EL.ipynb <==
import pandas as pd

RSV = pd.read_csv("06_RSV_palivo_EL.csv", sep=",", on_bad_lines="skip", quoting=0, encoding="utf-8", low_memory=False)

RSV.head(20)

RSV["Tovární značka"].value_counts()["TESLA"]

RSV.loc[RSV["Kategorie vozidla"] != "M1"]

RSV["Plně elektrické vozidlo"].unique()

import pandas as pd

df = pd.read_csv("06_RSV_palivo_EL.csv")

df_necisla = df[~df['PČV'].astype(str).str.isnumeric()]

print(df_necisla['PČV'])

pocet = df[df['PČV'].astype(str).str.contains('PROVOZOVANÉ', na=False)].shape[0]
print(pocet)
