==> 07_RSV_vlastnik_final.ipynb <==
import pandas as pd

RSV = pd.read_csv(
    "07_RSV_vlastnik_final.csv",nrows=8000000,sep=";",
    on_bad_lines="skip", quoting=0, encoding="utf-8",low_memory=False)

RSV.head(10)

#Typ subjektu 1 FO s trvalým v CR, 2 PO nebo podnikající FO 3 neztotožněný (zahraniční osoby, historická data)
#Vztah k vozidlu: 1-vlastník, 2-provozovatel

##### všechny hodnoty v Datum do přejmenovat na NaN a převést do stávajícího 07_RSV_vlastnik_final.csv

import pandas as pd
import numpy as np

df = pd.read_csv("07_RSV_nevalidni_datum_do.csv", sep=";", encoding="utf-8-sig")

# Přepiš celý sloupec 'Datum do' na NaN
df["Datum do"] = np.nan

# Ulož zpět
df.to_csv("07_RSV_nevalidni_datum_do.csv", sep=";", index=False, encoding="utf-8-sig")

print("Všechny hodnoty ve sloupci 'Datum do' byly přepsány na NaN.")

import pandas as pd

RSV = pd.read_csv(
    "07_RSV_nevalidni_datum_do.csv",sep=";",
    on_bad_lines="skip", quoting=0, encoding="utf-8",low_memory=False)

RSV["Datum do"].unique()

import pandas as pd

opraveny_df = pd.read_csv("07_RSV_nevalidni_datum_do.csv", sep=";", encoding="utf-8-sig")

final_df = pd.read_csv("07_RSV_vlastnik_final.csv", sep=";", encoding="utf-8-sig")

# Připoj opravené řádky na konec
spojene_df = pd.concat([final_df, opraveny_df], ignore_index=True)

spojene_df.to_csv("07_RSV_vlastnik_final.csv", sep=";", index=False, encoding="utf-8-sig")

print("Opravené řádky úspěšně přidány do 07_RSV_vlastnik_final.csv")

#### to samé u Datum od

import pandas as pd
import numpy as np

df = pd.read_csv("07_RSV_nevalidni_datum_od.csv", sep=";", encoding="utf-8-sig")

# Přepiš celý sloupec 'Datum do' na NaN
df["Datum od"] = np.nan

# Ulož zpět
df.to_csv("07_RSV_nevalidni_datum_od.csv", sep=";", index=False, encoding="utf-8-sig")

print("Všechny hodnoty ve sloupci 'Datum od' byly přepsány na NaN.")

import pandas as pd

RSV = pd.read_csv(
    "07_RSV_nevalidni_datum_od.csv",sep=";",
    on_bad_lines="skip", quoting=0, encoding="utf-8",low_memory=False)

RSV["Datum od"].unique()

import pandas as pd

# Načti opravený soubor s Datum od = NaN
opraveny_df = pd.read_csv("07_RSV_nevalidni_datum_od.csv", sep=";", encoding="utf-8-sig")

# Připoj na konec souboru bez načítání původního
opraveny_df.to_csv("07_RSV_vlastnik_final.csv", sep=";", index=False, encoding="utf-8-sig", mode="a", header=False)

print("Řádky úspěšně připojeny k 07_RSV_vlastnik_final.csv")
